refactor: combine SearchForm components
