--- src/Main/Header/SearchForm/CityIndex.js
@@ -1,10 +1,0 @@
-import styled from "styled-components";
-
-export default styled.span`
-  text-transform: uppercase;
-  line-height: 20px;
-  font-size: 16px;
-  color: #a0b0b9;
-  margin-left: auto;
-  margin-right: 12px;
-`;
