--- src/Main/Header/SearchForm/CityInput.js
@@ -1,5 +1,0 @@
-import styled from "styled-components";
-
-export default styled.input`
-  width: 100%;
-`;
