--- src/Main/Header/SearchForm/Dates/DateText.js
@@ -1,6 +1,0 @@
-import styled from "styled-components";
-
-export default styled.p`
-  margin: 0;
-  color: ${props => (props.disabled ? "#A0B0B9;" : "#4A4A4A")};
-`;
